Customer layers

+++ app/MongoDB/Data/repository/table_functions.py
@@ -7,7 +7,8 @@ def get_all_assets(mongo_object):
 
 
 def get_asset_by_id(mongo_object, object_id):
-    return mongo_object.find(_id=ObjectId(object_id)).first_or_none()
+    if ObjectId.is_valid(object_id):
+        return mongo_object.find(_id=ObjectId(object_id)).first_or_none()
 
 
 def get_assets_by_columnvalue(mongo_object, column_name, value):
@@ -20,6 +21,9 @@ def get_assets_by_columnvalue(mongo_object, column_name, value):
     elif isinstance(data_type, dict):
         return []
     else:
+        if ObjectId.is_valid(value):
+            value = ObjectId(value)
+
         return mongo_object.find(**{
             column_name: {
                 "$regex": value
@@ -28,12 +32,16 @@ def get_assets_by_columnvalue(mongo_object, column_name, value):
 
 
 def get_columns(mongo_object):
-    return [key for key in mongo_object.find().first_or_none().__dict__]
+    dummy_request = mongo_object.find().first_or_none()
+    if dummy_request is not None:
+        return [key for key in dummy_request.__dict__]
+    else:
+        return []
 
 
 def update_asset_by_column(mongo_object, column_name, value):
-    if column_name in mongo_object:
-        mongo_object[column_name] = value
+    if column_name in mongo_object.__dict__:
+        setattr(mongo_object, column_name, value)
     mongo_object.save()
     return mongo_object
 
@@ -43,7 +51,8 @@ def add_row(mongo_object, insert_dict):
 
 
 def drop_row_by_id(mongo_object, object_id):
-    return mongo_object.delete_one(_id=ObjectId(object_id))
+    if ObjectId.is_valid(object_id):
+        return mongo_object.delete_one(_id=ObjectId(object_id))
 
 
 def refresh_row(mongo_object):
@@ -51,4 +60,7 @@ def refresh_row(mongo_object):
 
 
 if __name__ == "__main__":
-    print(get_columns(Customer))
+    asset = get_all_assets(Customer)[0]
+    update_asset_by_column(asset, "id", 2)
+    print(asset)
+
